Refactor state definitions out of state machine file
